restart ntpd when dhcpv6_ntp_server sysevent is set ( OFW-2659 )
ntpd daemon starts when sysevent "wan-status" becomes "started".
But in dual-stack mode, this sysevent is set when erouter0 gets
IPv4 address. Because of that, NTPv6 option from DHCPv6 is not
applied to ntpd.

As a fix, register a new sysevent dhcpv6_ntp_server in rdk-wanmanager
and restart ntpd when "dhcpv6_ntp_server" is set.

Signed-off-by: Cibil <cpankiras.contractor@libertyglobal.com>