dual mode: dont try to stop DHCPv4 client on IPv6 only mode and bridge mode ( OFW-4323 )
temp fix erouter0 gets erouter ipv4 in bridge mode
,when moving from erouter mode to bridge mode using config file

[Issue]
DUT gets erouter ipv4 in bridge mode ,when moving from erouter mode to bridge mode using config file
Bridge mode for wan manager is not supported yet.

[Description]
On the basis of last_erouter_mode syscfg value, we  are exiting from Wan Manager. But this value was not
updated at that time. This issue is seen after Factory Reset or switching mode from router to bridge or vice-versa.
This issue is not seen with dmcli commands for eroutermodecontrol. It is only seen with the boot config.

[Solution]
 For MV2+, don't start/stop udhcpc and dibbler client in bridge_mode.

[Testing Performed]
- Enable dualstack by boot config file
- Check if Wan Manager is in ps output.
- Check erouter0 has IP4 and IPv6
- Enable bridge mode by boot config file
- Check if Wan Manager is in ps output.
- Check erouter0 has no IP4
- connected client get the public ip

Note: Need to be revisited once wanmanager supports bridge mode.

Signed-off-by: Jyoti <jjyoti@libertyglobal.com>