hot-standby: initial implementation for MV2+ platform
[Issue]
Hot-standby policy (AUTOWAN + Failover) does not work for MV2+ as an out-of-the-box solution.

[Description]
If an interface tries to run its SM while another interface is already initialized,
SM init fails for the first interface

[Solution]
change ISM_Status parameter to per-interface.

[Testing Performed]
Should not impact the normal code flow device is configured as static DOCSIS WAN

Signed-off-by: Devin Mungan(dmungan@libertyglobal.com)