fix DDNS configuration is not persistent across reboots and it throws error in UI ( OFW-4686 F3896SI-2049 )
[DESCRIPTION]
DDNS configuration settings are not persistent across reboots and it throws error in UI .

[FIX DESCRIPTION]
In Dualstack mode cpe is initially gettig IPV6 address followed by IPV4, when we get IPV6 address we are setting wan-status as started .
change in wan-satus triggers the service_ddns executable and ip validation checks in service-ddns fails since we are not assigned with IPV4 address and writes missconfiguration error in syscfg.db.
As a fix we are setting wan-status as started after getting both IPV6 and IPV4 address in dual stack mode.

[TESTING PERFORMED]
1. Load the DUT with test build and ensure CM is operational
2. Enable DDNS page in UI using below parameter via ACS
    SPV:"Device.DynamicDNS.X_LGI-COM_Enable" : 1
3. Configure DDNS settings via UI and it should be saved successfully
example:
DDNS Service : No-IP.com
   Username: xxx
   Password: yyy
   Domain name : zzz.ddns.net
4. Check the DDNS configuration via ACS/UI and it should be "NO_ERROR"
    GPV : "Device.DynamicDNS.Client."
5. Perform reboot via ACS/UI
6. Check the DDNS page in UI
   Result : UI displayed the DDNS settings without any error

Upstream Candidate: Yes

Signed-off-by: arudas <arudas.ext@libertyglobal.com>