temp_fix: GPV of "Device.DHCPv6.Client.1.Status" doesn't return the correct value ( F3896SI-2181 OFW-4958 )
BUG DESCRIPTION:
The status variable was updating the value based on the "Enable" parameter value stored in config structure.
This "Enable" DM value is initialized only during boot up and it was "FALSE" everytime.
Because of that the "status" DM always throw "Disabled".

During the GPV of "Enable", it doesn't read from the stored value in the config structure. Rather
it checks if dibbler-client is running or not. So "Enable" parameter always returns proper value.

FIX DESCRIPTION:
During GPV of 'status' parameter, checks if dibbler-client is running or not.
Based on that the status parameter value is returned.

Note: This change needs to be reverted once the DHCP manager component is available in the system.

TESTING PERFORMED:
Tested below scenarios:
1. Check Device.DHCPv6.Client.1.Status when dibbler-client is running. Returns "Enabled".
2. Check the same DM after stopping dibbler-client. Returns "Disabled".

Note:
No functional impact seen even if Status DM returns wrong  value.

Signed-off-by: Keerthana V J(kvijayakumarjalajaku.contractor@libertyglobal.com)