temp_fix: MV3 is offline in ACS while having connectivity [OFW-4999/F5685SI-1558]
[Issue]
Device.ManagementServer.URL is empty
Issue is due to dhcp option did not have ACS url.
As per design CPE was rejecting the IP address offer.
Selfheal triggers and notice erouter0 does not have IPv4 address and triggers dhcp client restart.
In this case flow reaches the service_udhcpc and service_wan
udhcpc is triggered with docsis options and it does not process the ACS url
Device will have IP and customer does not face issue, but the operator loses the management control on the CPE.

[Solution]
Fix the selfheal script and trigger the event to wan-manager to handle the specific MV3 case

[Testing]
N/A

Note: This commit is required for the docsis platform with Wanmanager, we need to revist this in R1.7. For now restrict to MV3

Signed-off-by: Karthik Shanbhag(kshanbhag@libertyglobal.com)