fix DHCPv6 client restart event not recorded in kibana server ( OFW-5358 F5685SI-1848 )
[Issue]
DHCPv6 restart event is not updated in ODH server and SelfHeal.txt file, since it is not triggered from test and diagnostic.

[Description]
DHCPv6 restart is triggered from rdk-Wanmanager now. In the current Implementation we are not triggering it from SelfHeal,
so the marker is not uploaded.

[Solution]
Added an event marker in rdk-Wanmanager component where the DHCPv6 client restart takes place.

[Testing]
1. Upgrade the image with new change
2. Enable telemetry with DCM file using "EVT_SYS_SH_DHCPV6Client_restart_split" marker
   {
        "header": "EVT_SYS_SH_DHCPV6Client_restart_split",
        "content": "com.cisco.spvtg.ccsp.wanmanager",
        "type": "<event>",
        "pollingFrequency": "0"
    }
3. Wait for the telemetry DCM file to be downloaded, this can be performed by GPV of Device.X_LGI-COM_Telemetry.
4. Wait for the telemetry logs to be pushed to ODH
5. Verify the marker and respective value in the telemetry log file on ODH

Signed-off-by: Durga Prasad D <ddemali.ext@libertyglobal.com>