fix wan-manger fails to restart dibbler client for mg0/voip0 ( OFW-5646 F5685SI-1724 )
[Issue]
IPOE sends message to wan-manager to renew IPv6.
Wan-manger kills the existing client, but does not start back the dibbler client for mg0/voip0

[Solution]
State machine is unable to find the dibbler client for mg0/voip0 as interface name is not phrased.
Interface name in required for MV3

[Testing]
Loaded image, confirm IPOE is working as expected.

[Note]
This change is only for 1.7 Mv3, 2.0 needs rework in dibbler scripts.
The change will made generic across platform in 2.0

Target branch : rnl17

Signed-off-by: Karthik Shanbhag(kshanbhag@libertyglobal.com)