fix wan-manger fails to restart dibbler client for mg0/voip0 ( OFW-5646 F5685SI-1724 )
[Issue]
IPOE sends message to wan-manager to renew IPv6.
Wan-manger kills the existing client, but does not start back the dibbler client for mg0/voip0

[Solution]
State machine is unable to find the dibbler client for mg0/voip0 as interface name is not phrased.
Interface name in required for MV3

[Testing]
Loaded image, confirm IPOE is working as expected.

Signed-off-by: Karthik Shanbhag(kshanbhag@libertyglobal.com)