remove route for the WAN prefix received ( OFW-5659 )
[Issue]
fixing regression after the fix for the issue ipv6 connectivity is not
working from the client connected to downstream router

[Issue Description]
Ipv6 ping from a LAN client of downstream router to a static ipv6
address of LAN client which comes in a unleased prefix is successful.
Also ping to WAN network from the static IP of direct LAN client is
successful after the change done in firewall as part of OFW-5659

[Fix Description]
The route to ipv6 wan prefix is not required after including the support
of dibbler-server-notify.sh

[Testing Performed]
1. Connected a LAN client
2. Assigned a static IP to the LAN client out of the brlan0 prefix (
choose any one prefix from the 8 devided prefix)
3. Verified Ipv6 ping to WAN network is failing.
4. Verified Ipv6 connectivity from the clients of downstream router is
working fine.

Signed-off-by: BijeeshaVS <bvattakaithayilsudha.contractor@libertyglobal.com>