fix IPV6 Internet is not working for dslite mode ( OFW-5934 )
[Issue]
ra-lifetime is 0 therefore LAN client won't have default IPv6 route.

[Description]
ra-lifetime in zebra.conf is set as 0 if wan-status is not "started"
when opensync is enabled, this is not a problem in practice because
opensync triggers radv-restart after wan-status is set to "started"
but when opensync is disabled, we rely on rdk-wanmanager's radv-restart
which is triggered before wan-status is set to started.

[Solution]
move radvd-restart along with dhcps-restart and ipv6_connection_status
setting after wan-status is set to "started"

[Testing Performed]
1. set eroutermode to DSLite
2. disable opensync
3. reboot device and confirm ra-lifetime is nonzero (/var/zebra.conf)

Signed-off-by: Devin Mungan(dmungan.ext@libertyglobal.com)