fix dml "Device.X_LGI-COM_WanManager.Interface.2.Status" doesn't get updated when Secondary WAN is not active ( OFW-5818 F3896SI-2661 )
[Issue Description]
While switching from enable to disable the Secondary WAN , the Status of secondary WAN is supposed to show as "Not Used".
But it still stays as "Standby" which is not expected.
The secondary WAN selection.Status is not being modified by WAN State Machine.

[Solution]
Modified code to return "NOT_SELECTED" when Secondary WAN is not enabled.

[Testing Performed]
MBU is connected to eth0 and enabled secondary WAN. The status is "Standy"
MBU is connected to eth0, secondary WAN enabled and disconnected RF. The status moved to "Active"
MBU is connected to eth0, Secondary WAN enabled and reconnected RF. The status moved to "Standy"
MBU is connected to eth0 and disabled Secondary WAN. The status is "Not Used"
MBU is disconnected from eth0, Secondary WAN is enabled and RF connected. The status moves to "Not Used"

Signed-off-by: Keerthana V J(kvijayakumarjalajaku.contractor@libertyglobal.com)