fix IPv6 issue on boot due to missing link-local address ( OFW-6061 F3896SI-2820 )
[Issue Description]
When WanManager tries to restart the dibbler-client, the check for erouter0 link-local address is done.
If the link-local address is not present , then dibbler-client will not get started.
The Monitor mechanism in WanManager does't catch it as restarts dibbler-client only if the dibbler-client was alredy started
by WanManager ISM before.
SO by any chance if the dibbler-client has not started even once , then there is no re-attempt to start it.
CPE will not have V6 connection.

[Fix Identified]
Added a check to start the dibbler if it doesn't started at all from the WanManager Monitor mechanism.

[Testing Performed]
Tested in the Ziggo environment by switching from R2.0 image to the fix image. The dibbler is running and device has
V6 connectivity.

Signed-off-by: Keerthana V J(kvijayakumarjalajaku.contractor@libertyglobal.com)