fix eRouter IPv4 address is empty message is showing in network log page ( OFW-6066 F3896SI-2855 F5685SI-2326 )
[Issue]
In CH MVlan, eRouter IPv4 address is empty message is showing in network log page.

[Description]
We are using the ifconfig utility to assign an IP to erouter0. During the
initial assignment, when the wanmanager receives data from service_udhcpc
to assign an IP, an RTM_DELADDR event is generated if the IP is assigned
using ifconfig.

[Solution]
Added a check to validate whether an IP address has been assigned or not after the AddrMon is triggered.

[Testing]
Verified that the "eRouter IPv4 address is empty" log message no longer appears in the network log.

[Upstream Candidate]
yes

Signed-off-by: Govindaraja N B <gnb.ext@libertyglobal.com>