set radvd_restart sysevent on IPv6 renewal to regenerate zebra.conf ( OFW-6407 )
[Issue]
[MV3-2.2] Internet is not working from LAN client via v6 Address while killed the eRouter global IPv6 address manually.

[Description]
When the IPv6 address is removed or the Dibbler client process is terminated, the Router Lifetime value becomes 0, causing routes to be deleted on LAN clients. This results in a loss of IPv6 connectivity.

[Solution]
Set radvd_restart sysevent on every IPv6 renewal to regenerate zebra.conf and restart zebra, ensuring updated RA messages are sent.

[Testing]
Verified the ipv6 internet connectivity.

[Upstream Candidate]
may be

Signed-off-by: Govindaraja N B <gnb.ext@libertyglobal.com>