github_migration - fix L3 network implementation introduced after wanmgr github migration
For some reason, adding IPv4 to the interface is moved from VISM to
dhcpv4 event handler. This means all interfaces with udhcpc running
will be assigned IPs as soon as the IP is leased. This implementation
causes problems when a secondary interface is configured in another
interface group.

Also, teardownIPv4 function no longer deletes the interface's IP.
Instead, wan_transition_ipv4_down function deletes it, but this
function is not called for standby interfaces, therefore for IPv4
deletion too, I'll revert back to the old solution.