prevent eth0 traffic leaking in brlan0 ( OFW-6612 CONNBE-331 )
[Issue]
eth0 traffic leaking into brlan0.

[Solution]
Enable packet forwarding on the WanOE eth interface when the
interface is part of the bridge — i.e., when the MBU feature is disabled.

Disable packet forwarding when the MBU feature is enabled but inactive —
i.e., when DOCSIS is active.

[Testing Performed]
Verify that packets do not leak from the eth interface to the bridge when:
- MBU feature is disabled
- MBU feature is enabled but DOCSIS is active

Signed-off-by: Devin Mungan(dmungan.ext@libertyglobal.com)