ignore IPv4 address deletion events from wanmanager ( OFW-6782 )
[Issue]
The AddrMon feature for IPv4 keeps triggering when udhcpc
is killed manually. After udhcpc is terminated, wanmanager
deletes the IPv4 on erouter0 which triggers a netlink
event, which causes udhcpc to restart.
This results in an infinite restart loop.

[Solution]
IPv4 address deletion events should be ignored when
they are triggered by wanmanager during IP address removal.

[Testing Performed]
1. Kill udhcpc on erouter0 manually.
2. Confirm IPv4 address recovery on erouter0.
3. Verify that udhcpc does not enter a restart loop.

Signed-off-by: Devin Mungan(dmungan.ext@libertyglobal.com)