Revert "fix internet is working in Ethernet-1 port in fallback scenario when Secondary WAN is active ( OFW-5927 F3896SI-2719 )"
This reverts commit 87636d4e991118f9f1b80de65852e67bb91823d7

Fix internet is working in Ethernet-1 port in fallback scenario when Secondary WAN is active ( OFW-5927 F3896SI-2719 )

[Issue]
Disconnecting MBU triggers eth0 port addition to brlan0

[Solution]
Change policy SM flow: if WAN is down, also check if the interface
is enabled or not. If disabled, trigger a re-select which eventually
will invalidate disabled interface and put eth0 back into the bridge.

[Testing Performed]
eth0 should be inside brlan0 when:
 - MBU feature is disabled (Device.X_LGI-COM_WanManager.Interface.2.Selection.Enable = FALSE)
eth0 should be outside brlan0 when:
 - MBU feature is enabled (Device.X_LGI-COM_WanManager.Interface.2.Selection.Enable = TRUE)
 - MBU is connected to MV2+.
 - MBU is not connected to MV2+.

Signed-off-by: Devin Mungan(dmungan.ext@libertyglobal.com)