fix DNS server parsing for multiple dns server in WanMgr_DmlDhcpcGetInfo ( OFW-7109 )
[Issue]
[MV3] [MV2PLUS] GPV of DNSServers dml showing incorrect values.

[Description]
On OFW platforms, we are using single space separated string to store
all DNS servers. Current code was using inet_addr() directly, so only
the first server was read. The second server was left empty and
showed as 255.255.255.255.

[Solution]
Changed code to split the DNS string and fill DNSServers[0] and DNSServers[1].
NumDnsServers is updated based on how many servers are found.

[Testing]
- Verified GPV of DNSServers dml shows both DNS servers correctly.

[Upstream Candidate]
No

Signed-off-by: Govindaraja N B <gnb.ext@libertyglobal.com>