fix default route missing when multiple RAs received on erouter0 ( OFW-7176 F3896SI-3542 )
[BUG DESCRIPTION]
Receiving multiple RA packets in response to a single RS on erouter0
can lead to multiple default routes being added to the gateway's
routing table.

The netmonitor process assumes only a single default route and uses a
flag (gw_v6_flag) to alternate between sending ADDROUTE and DELROUTE
notifications.

If two default routes are added and one is later removed, a
notification is correctly sent to the wanmanager. When the second
route is removed, netmonitor suppresses the notification due to the
state of gw_v6_flag and as a result wanmanager incorrectly assumes
that a default route still exists.

[FIX DESCRIPTION]
The fix removes the usage of the gw_v6_flag, allowing each default
route addition or removal to be processed individually.

[TESTING PERFORMED]
Step (1) Bring up the modem in dual stack/dslite mode.
Step (2) Execute dhcpv6_pd_60 in CD-Router.
Step (3) Verify default route is restored during prefix changes.

Signed-off-by: Andy Carvalho <acarvalho.contractor@libertyglobal.com>